Distance Between Colors

Calculation of the distances between the colors in the input palette. It also allows for calculation of the distances between the colors in a simulations of the color vision deficiency - deuteranopia, protanopia, and tritanopia.

palette_dist(x, cvd = NULL, severity = 1, metric = 2000)

Arguments

x
A vector of hexadecimal color descriptions

cvd
A type of color vision deficiency (CVD): "pro" (protanomaly), "deu" (deutanomaly), or "tri" (tritanomaly)

severity
Severity of the color vision defect, a number between 0 and 1

metric
A vector of color metric specifiers. Valid values are '1976', '1994', and '2000' (default), which refer to the year the metric was recommended by the CIE

Value

A matrix of distances between the original input palette and a simulation of the selected color vision deficiency - deuteranopia, protanopia, and tritanopia

Examples

rainbow_pal = rainbow(n = 7)
rainbow_pal

#> [1] "#FF0000FF" "#FFDB00FF" "#49FF00FF" "#00FF92FF" "#0092FFFF" "#4900FFFF"
#> [7] "#FF00DBFF"

palette_dist(rainbow_pal)

#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]      [,7]
#> [1,]   NA 52.96503 85.06611 84.19100 52.61836 51.46782  39.46279
#> [2,]   NA       NA 29.34232 34.74051 65.68745 92.04784  80.83188
#> [3,]   NA       NA       NA 12.13226 69.12782 81.23889 107.63470
#> [4,]   NA       NA       NA       NA 56.87222 74.19228 103.11485
#> [5,]   NA       NA       NA       NA       NA 35.47760  44.20031
#> [6,]   NA       NA       NA       NA       NA       NA  29.94683
#> [7,]   NA       NA       NA       NA       NA       NA        NA

palette_dist(rainbow_pal, cvd = "deu")

#>      [,1]     [,2]      [,3]     [,4]     [,5]     [,6]     [,7]
#> [1,]   NA 51.39677 46.855266 42.46564 58.52225 61.19252 55.57508
#> [2,]   NA       NA  7.725825 24.53764 75.63546 91.56339 80.29220
#> [3,]   NA       NA        NA 22.32611 69.93759 84.60113 74.34244
#> [4,]   NA       NA        NA       NA 47.89305 62.25387 53.37699
#> [5,]   NA       NA        NA       NA       NA 19.47525 11.33209
#> [6,]   NA       NA        NA       NA       NA       NA 11.16315
#> [7,]   NA       NA        NA       NA       NA       NA       NA

x = rcartocolor::carto_pal(11, "Vivid")
palette_dist(x)

#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,]   NA 52.04523 45.55534 39.23473 49.36807 46.51876 13.35131 49.85537
#>  [2,]   NA       NA 40.72139 61.23713 25.88310 31.39848 57.69268 17.19956
#>  [3,]   NA       NA       NA 23.92105 47.23524 22.12769 37.65264 31.46811
#>  [4,]   NA       NA       NA       NA 75.06609 36.10532 25.61009 57.81051
#>  [5,]   NA       NA       NA       NA       NA 42.17333 58.54488 42.47789
#>  [6,]   NA       NA       NA       NA       NA       NA 42.99029 26.57736
#>  [7,]   NA       NA       NA       NA       NA       NA       NA 57.46444
#>  [8,]   NA       NA       NA       NA       NA       NA       NA       NA
#>  [9,]   NA       NA       NA       NA       NA       NA       NA       NA
#> [10,]   NA       NA       NA       NA       NA       NA       NA       NA
#> [11,]   NA       NA       NA       NA       NA       NA       NA       NA
#>           [,9]    [,10]    [,11]
#>  [1,] 55.63423 25.37443 28.70389
#>  [2,] 12.84607 39.85479 37.74876
#>  [3,] 48.18521 58.75860 19.09441
#>  [4,] 77.24506 59.66097 21.99773
#>  [5,] 19.96106 26.17096 40.25210
#>  [6,] 36.19023 54.84155 27.32743
#>  [7,] 63.70692 36.47136 24.75022
#>  [8,] 30.90223 48.02986 32.30888
#>  [9,]       NA 37.95881 46.10770
#> [10,]       NA       NA 34.03568
#> [11,]       NA       NA       NA

palette_dist(x, cvd = "pro", severity = 1)

#>       [,1]     [,2]     [,3]     [,4]      [,5]     [,6]      [,7]      [,8]
#>  [1,]   NA 52.48196 25.11561 14.77268 50.197647 27.73398  8.180249 50.211156
#>  [2,]   NA       NA 35.44649 57.93880  2.784295 23.31238 55.779740  8.708812
#>  [3,]   NA       NA       NA 22.72090 32.791477 22.57092 23.661929 30.299207
#>  [4,]   NA       NA       NA       NA 55.102581 35.95804  6.935274 53.568400
#>  [5,]   NA       NA       NA       NA        NA 22.20630 53.152056  6.168612
#>  [6,]   NA       NA       NA       NA        NA       NA 32.502060 24.535526
#>  [7,]   NA       NA       NA       NA        NA       NA        NA 52.270175
#>  [8,]   NA       NA       NA       NA        NA       NA        NA        NA
#>  [9,]   NA       NA       NA       NA        NA       NA        NA        NA
#> [10,]   NA       NA       NA       NA        NA       NA        NA        NA
#> [11,]   NA       NA       NA       NA        NA       NA        NA        NA
#>            [,9]    [,10]     [,11]
#>  [1,] 55.621129 19.95530 23.176527
#>  [2,]  7.311807 36.21265 36.166575
#>  [3,] 41.654967 21.03491  2.467907
#>  [4,] 63.015709 29.17376 21.698042
#>  [5,]  9.936065 34.61957 33.569262
#>  [6,] 24.740536 10.65344 21.386683
#>  [7,] 60.200475 25.32213 22.177921
#>  [8,] 16.501685 35.96684 31.484004
#>  [9,]        NA 38.00344 42.022920
#> [10,]        NA       NA 18.974359
#> [11,]        NA       NA        NA

palette_dist(x, cvd = "pro", severity = 0.2)

#>       [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,]   NA 52.20713 39.53495 29.46635 51.12549 41.78917 10.55805 49.17448
#>  [2,]   NA       NA 37.49313 62.02816 19.78175 28.31784 57.19803 13.35525
#>  [3,]   NA       NA       NA 25.69001 39.24148 22.28508 34.87817 30.31760
#>  [4,]   NA       NA       NA       NA 68.07764 37.75698 19.46303 59.18465
#>  [5,]   NA       NA       NA       NA       NA 34.13492 57.79657 30.94296
#>  [6,]   NA       NA       NA       NA       NA       NA 41.45967 25.80445
#>  [7,]   NA       NA       NA       NA       NA       NA       NA 51.47508
#>  [8,]   NA       NA       NA       NA       NA       NA       NA       NA
#>  [9,]   NA       NA       NA       NA       NA       NA       NA       NA
#> [10,]   NA       NA       NA       NA       NA       NA       NA       NA
#> [11,]   NA       NA       NA       NA       NA       NA       NA       NA
#>           [,9]    [,10]    [,11]
#>  [1,] 56.21115 24.49188 26.16532
#>  [2,] 10.00157 37.01528 37.82102
#>  [3,] 43.96795 47.48796 17.37075
#>  [4,] 71.89308 47.51948 21.05439
#>  [5,] 16.85920 28.41884 38.52096
#>  [6,] 31.15593 44.22339 27.13749
#>  [7,] 62.79654 32.69802 23.68717
#>  [8,] 23.68762 41.28681 33.35172
#>  [9,]       NA 36.72867 45.12635
#> [10,]       NA       NA 30.16259
#> [11,]       NA       NA       NA